requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.3.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
eval_type_backport==0.2.0
executing==2.1.0
fastapi==0.103.1
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
idna==3.10
imageio==2.35.1
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.7
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4
numpy==2.1.2
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qudida==0.0.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.4.1
torchmetrics==1.4.3
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.23.2
wcwidth==0.2.13
yarl==1.14.0
